requirements.txt
absl-py==0.12.0
aiohttp==3.7.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
coverage==5.5
cryptography==3.4.7
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.29.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.7
jupyter-core==4.6.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.1.0
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pandas==1.2.4
parso==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==19.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
seaborn==0.11.1
setuptools==52.0.0.post20210125
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
